Change ID's from Integer to Long

+++ energymanagementsystem/src/main/java/rs/energymanagementsystem/energymanagementsystem/entities/GasHistoryData.java
@@ -17,7 +17,7 @@ public class GasHistoryData {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private Integer histDataId;
+    private Long histDataId;
 
     @Column
     private Integer gasConsumption;
